+++ pkg/adapter/config/interfaces.go
@@ -5,13 +5,6 @@ import (
 	v1 "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
-type etcdStore interface {
-	registry.Registry
-}
-type globalConfigStore interface {
-	registry.ConfigurationContext
-}
-
 type secret interface {
 	v1.SecretInterface
 }
@@ -20,6 +13,14 @@ type etcdRegistry interface {
 	registry.Registry
 }
 
+//nolint:unused
+//goland:noinspection GoUnusedType
+type globalConfigStore interface {
+	registry.ConfigurationContext
+}
+
+//nolint:unused
+//goland:noinspection GoUnusedType
 type configurationContext interface {
 	registry.ConfigurationContext
 }