+++ CHANGELOG.md
@@ -38,3 +38,4 @@ Config-specific:
 - [#39] Encrypt sensitive data
 - [#36] Set registry configuration for dogu and global config.
 - [#38] Censor all sensitive configuration data after applying the blueprint
+- [#45] Set registry configuration for encrypted values.